+++ examples/golang-client/.wundergraph/wundergraph.config.ts
@@ -1,0 +1,47 @@
+import { Application, configureWunderGraphApplication, cors, EnvironmentVariable, introspect } from '@wundergraph/sdk';
+import { golangClient } from '@wundergraph/golang-client';
+import server from './wundergraph.server';
+import operations from './wundergraph.operations';
+
+const countries = introspect.graphql({
+	apiNamespace: 'countries',
+	url: 'https://countries.trevorblades.com/',
+});
+
+const myApplication = new Application({
+	name: 'app',
+	apis: [countries],
+});
+
+// configureWunderGraph emits the configuration
+configureWunderGraphApplication({
+	application: myApplication,
+	server,
+	operations,
+	codeGenerators: [
+		{
+			templates: [
+				...golangClient.all({
+					packageName: 'client',
+				}),
+			],
+			path: '../go/pkg/client',
+		},
+	],
+	cors: {
+		...cors.allowAll,
+		allowedOrigins:
+			process.env.NODE_ENV === 'production'
+				? [
+						// change this before deploying to production to the actual domain where you're deploying your app
+						'http://localhost:3000',
+				  ]
+				: ['http://localhost:3000', new EnvironmentVariable('WG_ALLOWED_ORIGIN')],
+	},
+	dotGraphQLConfig: {
+		hasDotWunderGraphDirectory: false,
+	},
+	security: {
+		enableGraphQLEndpoint: process.env.NODE_ENV !== 'production',
+	},
+});