+++ packages/sdk/src/codegen/templates/typescript/jsonschema.ts
@@ -1,4 +1,4 @@
-import { Template, TemplateOutputFile } from '../../index';
+import { doNotEditHeader, Template, TemplateOutputFile } from '../../index';
 import { ResolvedWunderGraphConfig } from '../../../configure';
 import { formatTypeScript, loadFile } from './index';
 import Handlebars from 'handlebars';
@@ -19,7 +19,7 @@ export class JsonSchema implements Template {
 			{
 				path: 'jsonschema.ts',
 				content: formatTypeScript(content),
-				doNotEditHeader: true,
+				header: doNotEditHeader,
 			},
 		]);
 	}