+++ packages/nextjs/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.14](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.1.13...@wundergraph/nextjs@0.1.14) (2022-09-09)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.1.13](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.1.12...@wundergraph/nextjs@0.1.13) (2022-09-08)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.1.12](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.1.11...@wundergraph/nextjs@0.1.12) (2022-09-08)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
 ## [0.1.11](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.1.10...@wundergraph/nextjs@0.1.11) (2022-09-06)
 
 **Note:** Version bump only for package @wundergraph/nextjs